Extracted searchform component
