+++ test/pageObjects/pages/Home.page.js
@@ -1,36 +1,16 @@
 const Page = require("./page");
+const SearchForm = require("../components/SearchForm")
 
 class HomePage extends Page {
+  get searchForm() {
+    return new SearchForm();
+  }
   get acceptCookiesFrame() {
     return $("//iframe[@class='truste_popframe']");
   }
   get acceptCookiesButton() {
     return $(".call");
-  }
-  get inputForDestinationCity() {
-    return $("jb-autocomplete[data-qaid='autocomplete-city-selector'] input#jb-autocomplete-2-search");
-  }
-  get destinationCitySuggestedOption(){
-    return $("jb-autocomplete[data-qaid='autocomplete-city-selector'] strong");
-  }
-  get inputForDepartureDate() {
-    return $("#jb-date-picker-input-id-0");
-  }
-  get inputForReturnDate() {
-    return $("#jb-date-picker-input-id-1");
-  }
-  get searchButton() {
-    return $(".jb-booker-air-submit");
-  }
-  get errorForEmptyDestinationCity() {
-    return $("#jb-error-id-4");
   }   
-  get errorForEmptyDepartureDate() {
-    return $("#jb-error-id-0");
-  } 
-  get errorForEmptyReturnDate() {
-    return $("#jb-error-id-1");
-  } 
   get travelersDropdownTrigger() {
     return $('button[title="Travelers"]');
   } 
@@ -45,8 +25,7 @@ class HomePage extends Page {
   }
   get adultPassengersCounter() {
     return this.adultPassengersCalculator.$('span');
-  }
-  
+  }  
   
   async acceptCookies() {
     const acceptCookiesFrame = await this.acceptCookiesFrame;
